Feature/v1.5.0 (#87)
* v1.4.4 faster circleci testing

* docs fix: betas or beta allowed

* mouse probes have duplicate names, breaking dye-bias step, so it will fallback to linear-dye when dupes are present

* fixed bug, so if idat probes are missing/duplicate, dye switches to linear without breaking

* added more mouse array test coverage

* mouse unit test sample

* fixed epic+ qualityMask

* draft progress fixing mouse manifest and duplicate probes

* pulling v1.4.6 from master into v1.4.5 mouse support (which is partially complete) (#85)

* Feature/v1.4.4 (#83)

* v1.4.4 faster circleci testing

* docs fix: betas or beta allowed

* mouse probes have duplicate names, breaking dye-bias step, so it will fallback to linear-dye when dupes are present

* fixed bug, so if idat probes are missing/duplicate, dye switches to linear without breaking

* added more mouse array test coverage

* Feature/v1.4.6 (#84)

* v1.4.4 faster circleci testing

* docs fix: betas or beta allowed

* mouse probes have duplicate names, breaking dye-bias step, so it will fallback to linear-dye when dupes are present

* fixed bug, so if idat probes are missing/duplicate, dye switches to linear without breaking

* added more mouse array test coverage

* mouse unit test sample

* fixed epic+ qualityMask

* working draft

* CSVs will include all probe values

* CSV output shows all probes

* part-way through refactoring SigSet to replace raw_dataset.py and drop a layer of depth/abstraction

* WIP SigSet to replace raw_dataset, meth_data, and merge into SDC; currently on noob step

* mostly working after refactor. pipeline runs for sesame passing tests, and runs for minfi (not all tests pass yet)

* passing most tests, from 11 fails to 4 fails now

* down to 2 failed tests, and mouse bad probes handling, after reverting yesterday's testes

* all but 1 test passes. but mouse processing still off

* all caught up, before fixing mouse

* mouse working too

* moved old class objects to raw_dataset.py and renamed meth_dataset to sigset.py; dropped unused stuff from __init__ for each folder

* tests passing; new unit test matches betas with open-sesame output to within 0.01

* tests working

* v1.5.0 tests passing, and fix NOOB bug that changes all unit test ref outout; it was not excluding poobah/qualityMask failed probes from NOOB oobG/R in previous versions. Not sure if sesame matches this yet, but sesame does have a noMask() function in its NOOB that should align with this

* bug

* all tests pass with NOOB fix

* all unit tests passing; ref numbers updated to reflect the noob-poobah-qmask omission bug.

* unit tests passing but sesame not matching up at noob/dye stage

* testing against master prev version

* checkin back

* running large scale tests, looking for anything weird on large batches

* debugging: found GSM sample_ids are not picked up

* better error message when mix probe counts

* final edits; test then push

* tried to fix FTP error catching (no dice) and improved sesame comparison in mouse

* final tests should be passing on circleci

* typo

* typo

* missing sesame compare csvs for mouse test

* all tests passed on circleci, and I know the lingering discrepancy with sesame is likely a bug on their end in NOOB

* updated documentation; added '--no_quality_mask' to process CLI

* final edits to docs

+++ tests/processing/test_pipeline_batch_size.py
@@ -2,8 +2,9 @@ from methylprep.processing import pipeline
 import pandas as pd
 from pathlib import Path
 import numpy as np
+import unittest
 
-class TestBatchSize():
+class TestBatchSize(unittest.TestCase):
 
     """ TOO SLOW for CI - and redundant with test_pipeline
     @staticmethod
@@ -34,15 +35,26 @@ class TestBatchSize():
             raise AssertionError()
     """
 
-    @staticmethod
-    def test_with_batch_size():
+    def test_pipeline_sample_name_must_be_list(self):
         test_data_dir = 'docs/example_data/GSE69852'
-        df = pipeline.run_pipeline(test_data_dir, export=True, batch_size=1, sample_name='AdultLiver1')
+        with self.assertRaises(SystemExit):
+            df = pipeline.run_pipeline(test_data_dir, sample_name='AdultLiver1')
+
+    def test_with_batch_size(self):
+        test_data_dir = 'docs/example_data/GSE69852'
+        df = pipeline.run_pipeline(test_data_dir, export=True, batch_size=1, sample_name=['AdultLiver1'])
+        
+        ref_v145 = [
+            ['cg00063477',     4115.0,        172.0,        0.000,           0.0,       0.960,    4.580],
+            ['cg00121626',     3552.0,       3381.0,        0.000,           0.0,       0.512,    0.071],
+            ['cg27619353',     2204.0,       9713.0,        0.000,           0.0,       0.185,   -2.140],
+            ['cg27620176',     6052.0,         94.0,        0.001,           0.0,       0.985,    6.009],
+        ]
         ref = [
-            ['cg00063477',     4115.0,        172.0,        0.000,           1.0,       0.960,    4.580],
-            ['cg00121626',     3552.0,       3381.0,        0.000,           1.0,       0.512,    0.071],
-            ['cg27619353',     2204.0,       9713.0,        0.000,           1.0,       0.185,   -2.140],
-            ['cg27620176',     6052.0,         94.0,        0.001,           1.0,       0.985,    6.009],
+            ['cg00063477',     4125.0,        171.0,        0.000,           0.0,       0.960,    4.592],
+            ['cg00121626',     3562.0,       3391.0,        0.000,           0.0,       0.512,    0.071],
+            ['cg27619353',     2214.0,       9722.0,        0.000,           0.0,       0.185,   -2.135],
+            ['cg27620176',     6062.0,         90.0,        0.001,           0.0,       0.985,    6.074],
         ]
         ref_data = pd.DataFrame(ref, columns=['IlmnID','noob_meth','noob_unmeth','poobah_pval','quality_mask','beta_value','m_value']).set_index('IlmnID')
         data = df[0]._SampleDataContainer__data_frame.loc[ref_data.index]
@@ -61,15 +73,20 @@ class TestBatchSize():
             file.unlink()
 
 
-    @staticmethod
-    def test_batch_size_betas():
+    def test_batch_size_betas(self):
         test_data_dir = 'docs/example_data/GSE69852'
         betas = pipeline.run_pipeline(test_data_dir, betas=True, batch_size=1)
-        ref = [
+        ref_v145 = [
             ['cg00063477',    0.959879,           0.961307],
             ['cg00121626',    0.512332,           0.351993],
             ['cg27619353',    0.184946,           0.358009],
             ['cg27620176',    0.984706,           0.983877],
+        ]
+        ref = [ # v1.5.0+
+            ['cg00063477',    0.960196,           0.961819],
+            ['cg00121626',    0.512297,           0.352334],
+            ['cg27619353',    0.185489,           0.358258],
+            ['cg27620176',    0.985371,           0.984264],
         ]
         ref_data = pd.DataFrame(ref, columns=['IlmnID', '9247377093_R02C01','9247377085_R04C02']).set_index('IlmnID')
         test_betas = betas.loc[ref_data.index]
@@ -79,6 +96,9 @@ class TestBatchSize():
         #    raise AssertionError(f"{betas.iloc[0]['9247377093_R02C01']} != 0.23623395577166542")
         if not Path(test_data_dir, 'beta_values.pkl').is_file():
             raise AssertionError()
+        betas_pkl = pd.read_pickle(Path(test_data_dir, 'beta_values.pkl'))['9247377093_R02C01']
+        if not np.isclose(betas_pkl.loc[['cg00063477', 'cg00121626','cg27619353','cg27620176']], ref_data['9247377093_R02C01'], atol=0.001).all():
+            raise AssertionError("beta_values.pkl don't match expected test values")
         print(f"TEST OUTPUT FILES: {list(Path(test_data_dir).rglob('*'))}")
         for file in Path(test_data_dir).rglob('*.pkl'):
             file.unlink()
